+++ src/Slim/Middleware/IpAddressFilter.php
@@ -18,13 +18,19 @@ class IpAddressFilter extends Middleware
     /** @var  IpAddress */
     private $ipAddressMiddleware;
 
-    public function __construct($container, array $allowedIpAddresses)
+    // 選項：
+    // - checkProxyHeaders：允許從 x-forwarded-for 標頭中取得轉發的來源位址。注意：此 IP 可以偽造，不可用於有安全顧慮的場合！
+    //   trustedProxies：允許使用轉發位址的 proxies 位址。如果沒指定的話，就是接受任意 proxies。
+    public function __construct($container, array $allowedIpAddresses, array $options = [])
     {
         parent::__construct($container);
 
         $this->allowedIpAddresses = $allowedIpAddresses;
 
-        $this->ipAddressMiddleware = new IpAddress();
+        $checkProxyHeaders = $options['checkProxyHeaders'] ?? false;
+        $trustedProxies = $options['trustedProxies'] ?? [];
+
+        $this->ipAddressMiddleware = new IpAddress($checkProxyHeaders, $trustedProxies);
     }
 
     public function __invoke(Request $request, Response $response, callable $next)