+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Change Log
 ==========
 
+- IpAddressFilter 增加 options 參數
+    - 增加 checkProxyHeaders 與 trustedProxies，接受來自 proxy 的 forwarded 位址
+
 ### 0.4.0 (2017-07-31)
 
 - RouteLogger 增加 options 參數